backend/requirements.txt
# Backend Python dependencies
redis==4.5.1
pydantic==1.10.9
psycopg2-binary==2.9.7
python-dotenv==1.0.0
fastapi==0.95.0
uvicorn==0.23.2